JetStream pull consumer redesign (#115)
* JetStream pull consumer redesign

* Consume and fetch all

* Max bytes

* Header parser multi span bug fix

* Msg size fix

* Fixed tests and async sync bug

* Fixed tests

* Fixed tests

* Fixed tests

* Fixed tests

* Trace test issue

* Fixed warnings

* Removed flappy no-ACK test

This test was checking if the server was resending a message
which was not ACKed, after the ack_wait times out. Resending
unacknowledged messages is a server function we could test
under more controlled test setup and it's this part of the
test isn't testing NATS NET codebase. For some reason resend
function isn't working most of the time when run under the
GitHub Actions environment and creating false positives.
Hence I decided to remove this part of the test to avoid
unnecessary noise for the time being.

* End fetch with internal clock

This is a case where server request timeout might not
reach the client because of network anomalies.

* Default to fast ACK

* Options -> Opts refactor
* Ignore extra messages on consumer dispose
* Propagate JS options for ACK

* Test asserts

* Flappy test trace

* Public API fixes

* JSMsg hides the NatsMsg and proxy relevant fields
* Moved RawData to example
* JSContext NewInbox is now internal