Serialization API proposal cont.
This is the continuation of the previous proposal where
simplified the serialization APIs and made them compatible
with native AOT deployments.

With this change, we make it more obvious where to pass
the serializer into various calls. We also made the interface
generic to make it easier to implement in certain scenarios.