Async iterator subscription memory fix
The issue was happening with Perf tests hanging. It turns out when
subscription was handled within async iterator we loose a reference to
the subscription potentially because it's in a using as well. This
solution makes sure the connection object will keep a reference
so subscription object won't be collected because of the weak ref
in the subscription manager. Also we don't need to worry about
disposing the subscription now because as soon as the loop exits
it will be disposed.