Testing failground framework (#179)
* Failground tests

* JS publish reconnect and leader change fix

* Failground logging tweaks

* Failground logging tweaks

* Test flapper fixes

* JS consume idle heartbeat fix

Keep the idle heartbeat timer ticking so if all else fails, it will try to send
pull requests as long as we have an open connection. We also made a preventive
change to connection state management making it thread safe since we suspect
there may be race conditions on ARM processors for example.

* Test TLS first check update

Version check is available now server with TLS first is released.

* Merge fixes

* Stay connected test

* Ordered consumer tests

* Ordered consumer log fix

* Ordered consumer clean up on error

In case of an exception during consume we make sure to
delete the temporary consumer created.

* Must pass max-msgs

* TLS validation error logging on connection

When connection fails because of SSL validation checks the exception doesn't
tell you that (just says could not connect). If the user enables logging
then they'd at least be able to see if the connection failed because of
TLS validation errors and why.

* Ordered consumer sequence fix

Also added a memory test to make sure consumer subscription isn't
being GCed prematurely.

* Test JS cluster and reconnect utils

* Run the server strictly on single IP (127.0.0.1). This gives
  us predictable client URLs from the server, as opposed to all
  the IP of external interfaces, which might vary.
* Had to regenerate certs to have 127.0.0.1 in SAN fro the server.
* Made the cluster setup JetStream capable.
* Server can be restarted.