Fixed JetStream ACK custom serialization (#255)
* Fixed JetStream ACK custom serialization

When a custom serializer used JetStream msg ACKs are still being serialized
using the custom serializer which is resulting in +ACK being sent to the
server with a distorted format. Now we just used the NATS default serializer
on message ACKs to format the +ACK/NAK/etc. payload correctly.

ConsumerFetchTest.cs was also passing for the wrong reasons, which is also
fixed. Plus minor format issue fixed.

* Fixed consume ack test

* Test flap tuning