Update ConsumerCreateAction enum and related usage
ConsumerCreateAction enum has been updated to include a default value.
Previously nullable instances of this enum have been updated to
leverage the default value instead. The necessary adjustments have been
made in the related JSON serializer, consumer creation methods, and
tests to reflect this change.

When using nullable enum to accomodate for the absence of 'action' property
in JSON, we are forced to remove the type constraints on the serializer
class which is not ideal and leaves us to deal with null values.

Instead, having the default value of the enum as the first element lets
us using 'default' keyword instead of null in calls, making it less confusing.