Add repository rename announcement (#561)
Added an announcement about the upcoming rename of the current repository to `nats.net`. This change is reflected in a new file REPO_RENAME.md, detailing the rationale behind the renaming and its impact on developers. Also updated the README.md file to include a warning about the impending rename and a link to the announcement.