Add support for auth in URL (#670)
* Add support for auth in URL

* formatting

* undo not needed change

* address review

* use first uri to get auth information

* set username on multiple uris to the actual used one + better redaction

* Add tests and docs

* Add docs to the client

* Format

* Implement URL encoding for user credentials

* Refactor authentication precedence and URL redaction handling

Redact passwords and tokens at the `NatsUri` level to preserve the original
URIs, especially when used in WebSocket connections. This change aims to
maintain backward compatibility and prevent disruptions in WebSocket connections
that may rely on tokens or user-password pairs for proxy-level authentication.

* Remove unused test params

* Refactor redacted URI handling in NatsUri class

Modified the NatsUri class to store the redacted URI as a string instead of a
Uri object. This change simplifies the ToString method and ensures that sensitive
information like user credentials is consistently redacted for logging purposes.
Redacted string is created once in the ctor and avoiding Uri.ToString() calls
being made everytime it's needed.

* Move read URL credentials into NatsOpts

---------

Co-authored-by: Ziya Suzen <ziya@suzen.net>