Remove warning log on dropped message (#898)
Dropped messages can occur regularly. Depending on configuration / use case, this is even expected behavior. Logging a warning in the core NatsConnection potentially floods the logs with warnings and makes it difficult for clients to override. Logging on or otherwise handling dropped messages is better left for the user (i.e. by adding a custom MessageDropped event handler).